+++ test-helper/src/exe-sleep.rs
@@ -1,0 +1,4 @@
+fn main() {
+    std::thread::sleep(std::time::Duration::from_millis(2000));
+    println!("Have a good sleep.");
+}